build: configure multiple package exports
