--- packages/ffx-parser-ip/src/index.mts
@@ -1,1 +1,0 @@
-export { IP, IPv4, IPv6, parse } from "./lib/parser.mjs";
